movde db init and abstract discord helpers
