--- src/dbclient.py
@@ -1,4 +1,0 @@
-import boto3
-
-def get_dynamodb_resource():
-    return boto3.resource('dynamodb', region_name='us-east-2')
