+++ src/lambda_function.py
@@ -1,57 +1,32 @@
-from nacl.signing import VerifyKey
-from nacl.exceptions import BadSignatureError
-import constants
-import logger
+import boto3
 import bot
-import dbclient
+import constants
+import discord_helper
+import logger
 
-def verify_signature(event):
-    raw_body = event["rawBody"]
-    verify_key = VerifyKey(bytes.fromhex(constants.PUBLIC_KEY))
-    auth_sig = event["params"]["header"].get("x-signature-ed25519")
-    auth_ts  = event["params"]["header"].get("x-signature-timestamp")
-
-    try:
-        verify_key.verify(f"{auth_ts}{raw_body}".encode(), bytes.fromhex(auth_sig))
-    except BadSignatureError:
-        raise Exception("Verification failed")
-
-
-def is_lambda_request_event(event) -> bool:
-    if event["body-json"]:
-        return True
-    return False
-
-
-def is_ping_pong(body: dict) -> bool:
-    if body["type"]:
-        if body["type"] == 1:
-            return True
-    return False
+dynamodb = boto3.resource('dynamodb', region_name='us-east-2')
 
 def lambda_handler(event, context):
     print(f"event {event}") # debug print
 
     # verify the signature
     try:
-        verify_signature(event)
+        discord_helper.verify_signature(event)
     except Exception as e:
         raise Exception(f"[UNAUTHORIZED] Invalid request signature: {e}")
 
-    if not is_lambda_request_event(event):
+    if not event["body-json"]:
         return { "message": "Request is not Lambda event: 'body-json' not found" }
 
     body = event["body-json"]
 
-    if is_ping_pong(body): # Discord uses "ping pong" message to verify bot.
+    if discord_helper.is_ping_pong(body):
         print("is_ping_pong: True")
         response = constants.PING_PONG
     else:
         data = body["data"]
         command_name = data["name"]
         logger.log_command(command_name)
-
-        dynamodb = dbclient.get_dynamodb_resource()
 
         response = bot.process_bot_command(
             data=data,
